Put BigDecimal in bdmath
